+++ CHANGELOG.md
@@ -29,3 +29,4 @@ All notable changes to this project will be documented in this file.
 - Option for right aligned menu UI.
 - Stories and Areas sub menu in home screen.
 - Hide apply button if pattern viewer opened from home.
+- Reset the necessary global variables when exiting pattern state.